Crop treecar to car dimensions
Instead of using 32 x 32 pixel sprites, we will be cropping the sprites
to be as close to the vehicle body as possible. This ensures that the
wheels are placed in a sane place, rather than the outside edge of the
image, which could be much wider/longer than the actual vehicle.