Do not reset turning if opposite direction is down
When handling keyreleased events, only fire the endTurning events when
the opposite direction key is not currently being held down.

This resolves the issue where the car is "stuck" and unable to turn due
to a endTurning event resetting the angle/torque to zero while the
opposite direction is still being held.